permission controls

+++ posts/views.py
@@ -6,10 +6,13 @@ from .forms import PostForm
 from django.contrib import messages
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 from urllib.parse import quote
+from django.http import Http404
 
 
 
 def post_create(request):
+	if not (request.user.is_staff or request.user.is_superuser):
+		raise Http404
 	form = PostForm(request.POST or None, request.FILES or None)
 	if form.is_valid():
 		form.save()
@@ -21,6 +24,8 @@ def post_create(request):
 	return render(request, 'post_create.html', context)	
 
 def post_update(request, slug):
+	if not (request.user.is_staff or request.user.is_superuser):
+		raise Http404
 	post_object = get_object_or_404(Post, slug=slug)
 	form = PostForm(request.POST or None, request.FILES or None, instance=post_object)
 	if form.is_valid():
@@ -34,6 +39,8 @@ def post_update(request, slug):
 	return render(request,	'post_update.html', context)	
 
 def post_delete(request, slug):
+	if not (request.user.is_superuser):
+		raise Http404
 	post_object = get_object_or_404(Post, slug=slug)
 	post_object.delete()
 	messages.warning(request, "Post Deleted")
